server/requirements.txt
fastapi
uvicorn
torch
numpy
pillow
prometheus_client
python-multipart
